Get rid of Extension class, move settings into Session.
