+++ Sming/Components/ssl/Axtls/AxContext.cpp
@@ -65,8 +65,8 @@ Connection* AxContext::createClient()
 	assert(context != nullptr);
 
 	auto ssl_ext = ssl_ext_new();
-	ssl_ext_set_host_name(ssl_ext, session.extension.hostName.c_str());
-	ssl_ext_set_max_fragment_size(ssl_ext, session.extension.fragmentSize);
+	ssl_ext_set_host_name(ssl_ext, session.hostName.c_str());
+	ssl_ext_set_max_fragment_size(ssl_ext, session.fragmentSize);
 
 	auto id = session.getSessionId();
 	auto connection = new AxConnection(*this);
